Move some parser tests to fixtures
If we are going to have parsers in multiple languages, and we are, then we
should have an acceptance test suite that isn't specific to Go.

This commit begins the process of converting the Go-specific tests to
language-independent test fixtures.  Each test will be a workflow file,
annotated with one or more assertions about the result of parsing that
file: does it succeed or fail?  How many actions and workflows does it
produce?  What error messages does it emit, and at what severity?