Move some HCL-isms to a new fixture
The Go version of the parser allows strings in weird places, like
`"action" "a" ...` and `"on"="push"`.  The language spec doesn't allow
that, and neither does the WIP js parser.  I'm not going to change the Go
parser as part of this PR, but I want to identify HCL-isms that we didn't
intend to allow.  They're now in `hcl-subset-2.workflow`, which none of
the Go tests actually reads.

They could later move to the existing `hcl-subset.workflow` file once all
versions of the parser agree to reject them.