Allow commas in `env` blocks
language.md now admits that this form is legal:
```
env = {
  A = "b",
	C = "d"
}
```

Commas between entries in a hash/object are the norm in most languages.
They are allowed (optional) in HCL.  #36 introduced tests that assume that
this form is legal, and I believe there are at least a few main.workflow
files in the wild that use it.  30eeb74a0491 in #10 adapts the WIP JS
parser to allow them as well.

PR #40 proposes making the parser and fixtures stricter so they don't
allow these commas, but I think it's better to just make them an official
(rather than de facto) part of the language.  That's what this commit
does.